+++ CHANGES.rst
@@ -2,6 +2,16 @@
 Changelog
 =========
 
+Version 1.1
+-----------
+
+Unreleased
+
+* Support operating on multi-line records. (:issue:`1`)
+* Allow piping lines to an arbitrary command (``linesieve pipe``).
+* Allow hiding sections (``linesieve hide``).
+
+
 Version 1.0
 -----------